requirements.txt
aiohttp
requests
PyYAML